Camera: Set fps range according to HFR mode set.
Fps range will be set according to the HFR mode, for example if
HFR mode is 60, then fps range will be set as 60,60. This is
required to fix low fps observed in HFR modes.

Also, when we switch back to HFR off mode, we reset the fps range
in HAL to default fps setting. App can later change it if needed.

Change-Id: I131d9b31ff46a401bde36f4f43297de45d00a9f5